pyproject.toml
[project]
name = "vc-agent-marketplace"
version = "0.1.0"
description = "An intermediary platform connecting startup agents with VC agents for deal sourcing and communication"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "aiosqlite>=0.20.0",
    "anthropic>=0.40.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "run_server", "run_agent"]
packages = ["marketplace", "agents", "models"]
